Moved scrape UI to its own file
