+++ src/view/__init__.py
@@ -4,4 +4,5 @@ from .home_view_runelite import RuneLiteHomeView
 from .info_frame import InfoFrame
 from .output_log_frame import OutputLogFrame
 from .settings_view import SettingsView
+from .sprite_scraper_view import SpriteScraperView
 from .title_view import TitleView
